Add separate library index.ts
