+++ packages/tsargp/package.json
@@ -18,7 +18,7 @@
   },
   "exports": {
     ".": {
-      "types": "./dist/library/index.d.ts",
+      "types": "./dist/index.d.ts",
       "default": "./dist/index.min.js"
     },
     "./enums": {
@@ -60,8 +60,8 @@
   "license": "MIT",
   "scripts": {
     "prepare": "bun run build",
-    "build": "tsc -b && tsc -b src && bun build dist/library/index.js --outfile dist/index.min.js --minify && sed -i 's/const enum/enum/g' dist/enums/*.d.ts"
+    "build": "tsc -b && tsc -b src && bun build dist/index.js --outfile dist/index.min.js --minify && sed -i 's/const enum/enum/g' dist/enums/*.d.ts"
   },
   "type": "module",
-  "types": "./dist/library/index.d.ts"
+  "types": "./dist/index.d.ts"
 }
