+++ packages/tsargp/src/library/index.ts
@@ -1,22 +1,7 @@
-export type * from './options.js';
-export type * from './styles.js';
-
+export * from './options.js';
+export * from './styles.js';
 export * from './config.js';
 export * from './enums.js';
 export * from './formatter.js';
 export * from './parser.js';
 export * from './validator.js';
-
-export { allOf, oneOf, notOf, valuesFor } from './options.js';
-export {
-  seq,
-  style,
-  ext8,
-  rgb,
-  AnsiString,
-  AnsiMessage,
-  ErrorMessage,
-  JsonMessage,
-  TextMessage,
-  WarnMessage,
-} from './styles.js';
