+++ packages/tsargp/test/formatter/formatter.default.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'bun:test';
-import type { Options } from '../../src/library/options';
-import { format } from '../../src/library/formatter';
+import type { Options } from '../../src/library';
+import { format } from '../../src/library';
 
 describe('format', () => {
   describe('a default value is specified', () => {
