+++ packages/tsargp/test/parser/options/parser.command.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, jest } from 'bun:test';
-import { type Options } from '../../../src/library/options';
-import { parse } from '../../../src/library/parser';
+import type { Options } from '../../../src/library';
+import { parse } from '../../../src/library';
 
 process.env['FORCE_WIDTH'] = '0'; // omit styles
