+++ packages/tsargp/test/parser/parser.inline.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, jest } from 'bun:test';
-import { type Options } from '../../src/library/options';
-import { type ParsingFlags, parse } from '../../src/library/parser';
+import type { Options, ParsingFlags } from '../../src/library';
+import { parse } from '../../src/library';
 
 process.env['FORCE_WIDTH'] = '0'; // omit styles
